+++ tests/canopen/test_canopen_dictionary_v2.py
@@ -143,7 +143,6 @@ def test_registers_from_canopen_objects_have_object_reference():
     "register_uid, subnode, expected_monitoring_address",
     [("DRV_DIAG_ERROR_LAST", 1, 0x000F), ("DRV_DIAG_ERROR_LAST", 2, 0x000F)],
 )
-@pytest.mark.no_connection
 def test_register_monitoring_address(register_uid, subnode, expected_monitoring_address):
     dictionary_path = tests.resources.canopen.TEST_DICT_CAN_AXIS
     canopen_dict = CanopenDictionaryV2(dictionary_path)