+++ tests/ethercat/test_pdo_network_manager.py
@@ -23,7 +23,8 @@ def test_pdos_min_refresh_rate(net: "EthercatNetwork"):
 
 
 @pytest.mark.ethercat
-def test_pdos_watchdog_exception_auto(net: "EthercatNetwork", servo: "EthercatServo"):  # noqa: ARG001
+@pytest.mark.usefixtures("servo")
+def test_pdos_watchdog_exception_auto(net: "EthercatNetwork"):
     exceptions = []
 
     def exception_callback(exc):
@@ -36,11 +37,12 @@ def test_pdos_watchdog_exception_auto(net: "EthercatNetwork", servo: "EthercatSe
     net.pdo_manager.unsubscribe_to_exceptions(exception_callback)
     assert len(exceptions) > 0
     exception = exceptions[0]
-    assert str(exception) == "The sampling time is too high. The max sampling time is 3276.75 ms."
+    assert str(exception) == "The sampling time is too high. The max sampling time is 3.27675 s."
 
 
 @pytest.mark.ethercat
-def test_pdos_watchdog_exception_manual(net: "EthercatNetwork", servo: "EthercatServo"):  # noqa: ARG001
+@pytest.mark.usefixtures("servo")
+def test_pdos_watchdog_exception_manual(net: "EthercatNetwork"):
     exceptions = []
 
     def exception_callback(exc):
@@ -54,7 +56,7 @@ def test_pdos_watchdog_exception_manual(net: "EthercatNetwork", servo: "Ethercat
     assert len(exceptions) > 0
     exception = exceptions[0]
     assert (
-        str(exception) == "The watchdog timeout is too high. The max watchdog timeout is 6553.5 ms."
+        str(exception) == "The watchdog timeout is too high. The max watchdog timeout is 6.5535 s."
     )