+++ tests/ethernet/test_ethernet_dictionary_v2.py
@@ -128,7 +128,6 @@ def test_safety_pdo_not_implemented():
     "register_uid, subnode, expected_monitoring_address",
     [("DRV_DIAG_ERROR_LAST", 1, 0x000F), ("DRV_DIAG_ERROR_LAST", 2, 0x080F)],
 )
-@pytest.mark.no_connection
 def test_register_monitoring_address(register_uid, subnode, expected_monitoring_address):
     dictionary_path = tests.resources.ethernet.TEST_DICT_ETHERNET_AXIS
     ethernet_dict = EthernetDictionaryV2(dictionary_path)