+++ tests/test_servo.py
@@ -29,6 +29,7 @@ from ingenialink.exceptions import (
 )
 from ingenialink.register import RegAddressType
 from ingenialink.servo import Servo, ServoState
+from ingenialink.utils._utils import convert_bytes_to_dtype
 
 if TYPE_CHECKING:
     from summit_testing_framework.environment import Environment
@@ -79,8 +80,8 @@ class SDOReadTimeoutManager:
         self.__net._ecat_master.sdo_read_timeout = self.__initial_value
 
 
-@pytest.fixture()
-def create_monitoring(servo, net):
+@pytest.fixture
+def create_monitoring(servo):
     skip_if_monitoring_is_not_available(servo)
     servo.monitoring_disable()
     servo.monitoring_remove_all_mapped_registers()
@@ -91,12 +92,12 @@ def create_monitoring(servo, net):
     servo.write("MON_DIST_FREQ_DIV", divisor, subnode=0)
     servo.write("MON_CFG_SOC_TYPE", 0, subnode=0)
     servo.write("MON_CFG_WINDOW_SAMP", MONITORING_NUM_SAMPLES, subnode=0)
-    yield servo, net
+    yield servo
     servo.monitoring_disable()
 
 
 @pytest.fixture()
-def create_disturbance(servo, net):
+def create_disturbance(servo):
     skip_if_monitoring_is_not_available(servo)
     data = list(range(DISTURBANCE_NUM_SAMPLES))
     servo.disturbance_disable()
@@ -105,16 +106,14 @@ def create_disturbance(servo, net):
         channel=0, uid="CL_POS_SET_POINT_VALUE", size=DISTURBANCE_CH_DATA_SIZE
     )
     servo.disturbance_write_data(0, RegDtype.S32, data)
-    yield servo, net
+    yield servo
     servo.disturbance_disable()
 
 
 @pytest.mark.canopen
 @pytest.mark.ethernet
 @pytest.mark.ethercat
-def test_save_configuration(servo, net):
-    assert servo is not None and net is not None
-
+def test_save_configuration(servo) -> None:
     filename = "temp_config"
 
     servo.save_configuration(filename)
@@ -193,8 +192,7 @@ def test_check_configuration(virtual_drive):
 @pytest.mark.canopen
 @pytest.mark.ethernet
 @pytest.mark.ethercat
-def test_load_configuration(servo, net):
-    assert servo is not None and net is not None
+def test_load_configuration(servo) -> None:
     filename = "temp_config"
     servo.save_configuration(filename)
     assert os.path.isfile(filename)
@@ -218,7 +216,9 @@ def test_load_configuration_strict(mocker, virtual_drive_custom_dict):  # noqa: 
     dictionary = virtual_drive_resources.VIRTUAL_DRIVE_V2_XDF
     _, _, servo = virtual_drive_custom_dict(dictionary)
     test_file = tests.resources.TEST_CONFIG_FILE
-    mocker.patch("ingenialink.servo.Servo.write", side_effect=ILError("Error writing"))
+    mocker.patch(
+        "ingenialink.ethernet.servo.EthernetServo._write_raw", side_effect=ILError("Error writing")
+    )
     with pytest.raises(ILError) as exc_info:
         servo.load_configuration(test_file, strict=True)
     assert (
@@ -230,9 +230,7 @@ def test_load_configuration_strict(mocker, virtual_drive_custom_dict):  # noqa: 
 @pytest.mark.canopen
 @pytest.mark.ethernet
 @pytest.mark.ethercat
-def test_load_configuration_file_not_found(servo, net):
-    assert servo is not None and net is not None
-
+def test_load_configuration_file_not_found(servo) -> None:
     filename = "can_config.xdf"
     with pytest.raises(FileNotFoundError):
         servo.load_configuration(filename)
@@ -242,9 +240,7 @@ def test_load_configuration_file_not_found(servo, net):
 @pytest.mark.canopen
 @pytest.mark.ethernet
 @pytest.mark.ethercat
-def test_load_configuration_invalid_subnode(setup_descriptor, servo, net, subnode):
-    assert servo is not None and net is not None
-
+def test_load_configuration_invalid_subnode(setup_descriptor, servo, subnode) -> None:
     filename = setup_descriptor.config_file
     with pytest.raises(ValueError):
         servo.load_configuration(filename, subnode=subnode)
@@ -253,9 +249,7 @@ def test_load_configuration_invalid_subnode(setup_descriptor, servo, net, subnod
 @pytest.mark.canopen
 @pytest.mark.ethernet
 @pytest.mark.ethercat
-def test_load_configuration_to_subnode_zero(setup_descriptor, servo, net):
-    assert servo is not None and net is not None
-
+def test_load_configuration_to_subnode_zero(setup_descriptor, servo) -> None:
     path = setup_descriptor.config_file
     assert isinstance(path, Path)
     filename = path.as_posix()
@@ -375,9 +369,7 @@ def test_restore_safe_parameters(servo, environment: "Environment") -> None:
 @pytest.mark.canopen
 @pytest.mark.ethernet
 @pytest.mark.ethercat
-def test_read(servo, net):
-    assert servo is not None and net is not None
-
+def test_read(servo) -> None:
     value = servo.read("DRV_STATE_STATUS")
     assert value is not None
 
@@ -385,9 +377,7 @@ def test_read(servo, net):
 @pytest.mark.canopen
 @pytest.mark.ethernet
 @pytest.mark.ethercat
-def test_write(servo, net):
-    assert servo is not None and net is not None
-
+def test_write(servo) -> None:
     reg = "CL_AUX_FBK_SENSOR"
     value = 4
 
@@ -415,7 +405,7 @@ def test_monitoring_enable_disable(servo):
 @pytest.mark.canopen
 @pytest.mark.ethercat
 def test_monitoring_remove_data(create_monitoring):
-    servo, _ = create_monitoring
+    servo = create_monitoring
     servo.monitoring_enable()
     servo.write("MON_CMD_FORCE_TRIGGER", 1, subnode=0)
     assert servo.read("MON_CFG_BYTES_VALUE", subnode=0) > 0
@@ -450,7 +440,7 @@ def test_monitoring_map_register(servo):
 @pytest.mark.canopen
 @pytest.mark.ethercat
 def test_monitoring_data_size(create_monitoring):
-    servo, _ = create_monitoring
+    servo = create_monitoring
     servo.monitoring_enable()
     servo.write("MON_CMD_FORCE_TRIGGER", 1, subnode=0)
     assert servo.monitoring_get_bytes_per_block() == MONITORING_CH_DATA_SIZE
@@ -463,7 +453,7 @@ def test_monitoring_data_size(create_monitoring):
 @pytest.mark.canopen
 @pytest.mark.ethercat
 def test_monitoring_read_data(create_monitoring):
-    servo, _ = create_monitoring
+    servo = create_monitoring
     servo.monitoring_enable()
     servo.write("MON_CMD_FORCE_TRIGGER", 1, subnode=0)
     time.sleep(1)
@@ -492,7 +482,7 @@ def test_disturbance_enable_disable(servo):
 @pytest.mark.canopen
 @pytest.mark.ethercat
 def test_disturbance_remove_data(create_disturbance):
-    servo, _ = create_disturbance
+    servo = create_disturbance
     servo.disturbance_enable()
     assert (
         servo.read("DIST_CFG_BYTES", subnode=0)
@@ -529,7 +519,7 @@ def test_disturbance_map_register(servo):
 @pytest.mark.canopen
 @pytest.mark.ethercat
 def test_disturbance_data_size(create_disturbance):
-    servo, _ = create_disturbance
+    servo = create_disturbance
     servo.disturbance_enable()
     assert servo.disturbance_data_size == DISTURBANCE_CH_DATA_SIZE * DISTURBANCE_NUM_SAMPLES
     servo.disturbance_remove_data()
@@ -797,9 +787,12 @@ def test__adapt_configuration_file_storage_value(servo, uid, subnode, value, nod
     node_id_reg = servo.dictionary.registers(subnode)[uid]
     node_id_reg._CanopenRegister__is_node_id_dependent = dependent
     conf_file.add_register(node_id_reg, value)
-    adapted_register = servo._adapt_configuration_file_storage_value(
-        conf_file, conf_file.registers[0]
+    adapted_register_bytes = servo._adapt_configuration_file_storage_value(
+        configuration_file=conf_file,
+        config_register=conf_file.registers[0],
+        target_register=node_id_reg,
     )
+    adapted_register = convert_bytes_to_dtype(adapted_register_bytes, node_id_reg.dtype)
     if dependent:
         assert adapted_register == (value - node_id + servo.target)
     else: