+++ CHANGELOG.md
@@ -2,7 +2,11 @@
 
 ## [Unreleased]
 ### Added
-- Parsing of Dictionary tables
+- Add support for tables for XDF and XCF and CSV export. XCF version has been updated to 2.2
+- Added context-like behaviour to EtherCAT network
+
+### Fixed
+- Fix sdo timeout rollback issue during store and firmware loading operations.
 
 ## [7.5.2] - 2025-11-24
 ### Added